requirements.txt
streamlit
torch
ultralytics
Pillow
numpy
opencv-python-headless
